Fix compiling large files
	2016-03-15  Richard Henderson  <rth@redhat.com>
	* line-map.c (new_linemap): Make alloc_size a size_t.

+++ libcpp/line-map.c
@@ -364,7 +364,7 @@ new_linemap (struct line_maps *set,
   if (LINEMAPS_USED (set, macro_map_p) == LINEMAPS_ALLOCATED (set, macro_map_p))
     {
       /* We ran out of allocated line maps. Let's allocate more.  */
-      size_t alloc_size;
+      unsigned alloc_size;
 
       /* Cast away extern "C" from the type of xrealloc.  */
       line_map_realloc reallocator = (set->reallocator
